superscript.py v 0.0.6.002
