+++ data-analysis/data.py
@@ -34,6 +34,32 @@ def get_team_metrics_data(apikey, competition, team_num):
 	mdata = db.team_metrics
 	return mdata.find_one({"competition" : competition, "team": team_num})
 
+def get_match_data_formatted(apikey, competition):
+	client = pymongo.MongoClient(apikey)
+	db = client.data_scouting
+	mdata = db.teamlist
+	x=mdata.find_one({"competition":competition})
+	out = {}
+	for i in x:
+		try:
+			out[int(i)] = unkeyify_2l(get_team_match_data(apikey, competition, int(i)).transpose().to_dict())
+		except:
+			pass
+	return out
+
+def get_metrics_data_formatted(apikey, competition):
+	client = pymongo.MongoClient(apikey)
+	db = client.data_scouting
+	mdata = db.teamlist
+	x=mdata.find_one({"competition":competition})
+	out = {}
+	for i in x:
+		try:
+			out[int(i)] = d.get_team_metrics_data(apikey, competition, int(i))
+		except:
+			pass
+	return out
+
 def get_pit_data_formatted(apikey, competition):
 	client = pymongo.MongoClient(apikey)
 	db = client.data_scouting
@@ -45,6 +71,20 @@ def get_pit_data_formatted(apikey, competition):
 			out[int(i)] = get_team_pit_data(apikey, competition, int(i))
 		except:
 			pass
+	return out
+
+def get_pit_variable_data(apikey, competition):
+	client = pymongo.MongoClient(apikey)
+	db = client.data_processing
+	mdata = db.team_pit
+	out = {}
+	return mdata.find()
+
+def get_pit_variable_formatted(apikey, competition):
+	temp = get_pit_variable_data(apikey, competition)
+	out = {}
+	for i in temp:
+		out[i["variable"]] = i["data"]
 	return out
 
 def push_team_tests_data(apikey, competition, team_num, data, dbname = "data_processing", colname = "team_tests"):
@@ -86,17 +126,4 @@ def unkeyify_2l(layered_dict):
 			add.append([j,layered_dict[i][j]])
 		add.sort(key = lambda x: x[0])
 		out[i] = list(map(lambda x: x[1], add))
-	return out
-
-def get_match_data_formatted(apikey, competition):
-	client = pymongo.MongoClient(apikey)
-	db = client.data_scouting
-	mdata = db.teamlist
-	x=mdata.find_one({"competition":competition})
-	out = {}
-	for i in x:
-		try:
-			out[int(i)] = unkeyify_2l(get_team_match_data(apikey, competition, int(i)).transpose().to_dict())
-		except:
-			pass
 	return out
